# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Career Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs with filters
      parameters:
        - name: keyword
          in: query
          schema: { type: string }
        - name: company
          in: query
          schema: { type: string }
        - name: location
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: experience
          in: query
          schema: { type: string }
        - name: salaryMin
          in: query
          schema: { type: integer }
        - name: salaryMax
          in: query
          schema: { type: integer }
        - name: remote
          in: query
          schema: { type: boolean }
        - name: type
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 12 }
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"

    post:
      operationId: createJob
      tags: [admin]
      summary: Create a job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/featured:
    get:
      operationId: listFeaturedJobs
      tags: [jobs]
      summary: Get featured jobs for homepage
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Featured jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /jobs/trending:
    get:
      operationId: listTrendingJobs
      tags: [jobs]
      summary: Get trending/most-viewed jobs
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 8 }
      responses:
        "200":
          description: Trending jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /jobs/latest:
    get:
      operationId: listLatestJobs
      tags: [jobs]
      summary: Get latest job postings
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Latest jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found

    patch:
      operationId: updateJob
      tags: [admin]
      summary: Update a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

    delete:
      operationId: deleteJob
      tags: [admin]
      summary: Delete a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /jobs/{id}/view:
    post:
      operationId: recordJobView
      tags: [jobs]
      summary: Record a job view (for trending)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: View recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /stats/top-jobs:
    get:
      operationId: getTopViewedJobs
      tags: [stats]
      summary: Get most viewed jobs for admin analytics
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Top viewed jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
...
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [jobs]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        companyLogo:
          type: ["string", "null"]
        companyWebsite:
          type: ["string", "null"]
        location:
          type: string
        remote:
          type: boolean
        jobType:
          type: string
          description: "full-time, part-time, internship, contract, walkin, government"
        category:
          type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        salaryDisplay:
          type: ["string", "null"]
        experience:
          type: string
        education:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        selectionProcess:
          type: ["string", "null"]
        applyLink:
          type: string
        deadline:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        status:
          type: string
          description: "published, draft"
        views:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, slug, companyId, companyName, location, remote, jobType, category, experience, education, description, skills, applyLink, featured, status, views, createdAt, updatedAt]
    JobInput:
      type: object
      properties:
        title:
          type: string
        companyId:
          type: integer
        location:
          type: string
        remote:
          type: boolean
          default: false
        jobType:
          type: string
        category:
          type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        salaryDisplay:
          type: ["string", "null"]
        experience:
          type: string
        education:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        selectionProcess:
          type: ["string", "null"]
        applyLink:
          type: string
        deadline:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
          default: false
        status:
          type: string
          default: "draft"
      required: [title, companyId, location, jobType, category, experience, education, description, skills, applyLink]
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        companyId:
          type: integer
        location:
          type: string
        remote:
          type: boolean
        jobType:
          type: string
        category:
          type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        salaryDisplay:
          type: ["string", "null"]
        experience:
          type: string
        education:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        selectionProcess:
          type: ["string", "null"]
        applyLink:
          type: string
        deadline:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        status:
          type: string
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [jobs, total, page, limit, totalPages]
...
    NewsletterInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
      required: [email]