# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Career Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface JobUpdate {
  title?: string;
  companyId?: number;
  location?: string;
  remote?: boolean;
  jobType?: string;
  category?: string;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  /** @nullable */
  salaryDisplay?: string | null;
  experience?: string;
  education?: string;
  description?: string;
  skills?: string[];
  /** @nullable */
  selectionProcess?: string | null;
  applyLink?: string;
  /** @nullable */
  deadline?: string | null;
  tags?: string[];
  featured?: boolean;
  status?: string;
}